requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.1.1
cloudinary